adapter/callback: renamed client to callback and updated the tests
